Update server: cập nhật logic role và permission

+++ rently-server/src/routes/role/role.service.ts
@@ -10,7 +10,11 @@ import {
   isNotFoundPrismaError,
   isUniqueConstraintPrismaError,
 } from 'src/shared/helpers'
-import { RoleAlreadyExistsException } from 'src/routes/role/role.error'
+import {
+  ProhibitedActionOnBaseRoleException,
+  RoleAlreadyExistsException,
+} from 'src/routes/role/role.error'
+import { RoleName } from 'src/shared/constants/role.constant'
 
 @Injectable()
 export class RoleService {
@@ -60,12 +64,23 @@ export class RoleService {
     updatedById: number
   }) {
     try {
-      const role = await this.roleRepo.update({
+      const role = await this.roleRepo.findById(id)
+
+      if (!role) {
+        throw NotFoundRecordException
+      }
+
+      // không cho phép cập nhật role admin
+      if (role.name === RoleName.Admin) {
+        throw ProhibitedActionOnBaseRoleException
+      }
+
+      const updatedRole = await this.roleRepo.update({
         id,
         updatedById,
         data,
       })
-      return role
+      return updatedRole
     } catch (error) {
       if (isNotFoundPrismaError(error)) {
         throw NotFoundRecordException
@@ -82,12 +97,29 @@ export class RoleService {
 
   async delete({ id, deletedById }: { id: number; deletedById: number }) {
     try {
+      const role = await this.roleRepo.findById(id)
+
+      if (!role) {
+        throw NotFoundRecordException
+      }
+
+      // không cho phép xóa 3 role mặc định
+      const baseRole: string[] = [
+        RoleName.Admin,
+        RoleName.Client,
+        RoleName.Landlord,
+      ]
+      if (baseRole.includes(role.name)) {
+        throw ProhibitedActionOnBaseRoleException
+      }
+
       await this.roleRepo.delete({
         id,
         deletedById,
       })
+
       return {
-        message: 'Delete successfully',
+        message: 'Xoá role thành công',
       }
     } catch (error) {
       if (isNotFoundPrismaError(error)) {
